pyproject.toml
[project]
name = "subconcious"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "subconscious-sdk>=0.2.0",
    "mcp>=1.7.0",
    "httpx>=0.27.0",
]
